chemical_project/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
fonttools==4.60.1
gunicorn==23.0.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.5
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
